Handle empty JAVA_HOME in installation script (#66)
Previously the JAVA_HOME was evaluated when gppkg install the package.
If the JAVA_HOME is empty at that time, a error-formed PATH will be
injected into the greenplum_path.sh, like PATH=/bin:/usr/bin:xxxx.

This was wrong since user cannot easily fix the JAVA_HOME by just set it
in the script.

This patch postpone the VAR evaluation to the source stage. And reports
an warning message if the JAVA_HOME is empty.